Expose startServer
