+++ bin/cucumber-language-server.cjs
@@ -1,3 +1,10 @@
 #!/usr/bin/env node
 require('source-map-support').install()
-require('../dist/cjs/src/server')
+const startServer = require('../dist/cjs/src/startServer')
+
+const wasmUrls = {
+  java: `node_modules/@cucumber/language-service/tree-sitter-java.wasm`,
+  typescript: `node_modules/@cucumber/language-service/tree-sitter-typescript.wasm`,
+}
+
+startServer(wasmUrls)
