+++ src/index.ts
@@ -1,1 +1,1 @@
-export * from './CucumberLanguageServer.js'
+export * from './startServer.js'
